pyproject.toml
[tool.poetry]
name = "csppeleenet.pytorch"
version = "0.1.0"
description = ""
authors = ["tomotana14 <tomotana14@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
torch = "^1.6.0"
torchvision = "^0.7.0"
pytorch-ignite = "^0.4.1"
tqdm = "^4.49.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
